+++ srcs/utils/semaphore.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-semaphore::semaphore(unsigned long initialCount) : m_count(initialCount) {}
+semaphore::semaphore(uint32_t initialCount) : m_count(initialCount) {}
 
 semaphore::~semaphore() {};
 
@@ -15,7 +15,7 @@ void semaphore::notify() {
 void semaphore::wait() {
     std::unique_lock<decltype(m_mutex)> lock(m_mutex);
     if (!m_count)
-        m_condition.wait(lock);                                                    // TODO voir pour if !
+        m_condition.wait(lock);
     --m_count;
 }
